hosts-monitor-internal/requirements.txt
azure-common==1.1.28
azure-core==1.38.0
azure-identity==1.25.1
azure-mgmt-core==1.6.0
azure-mgmt-dns==9.0.0
azure-mgmt-privatedns==1.2.0
azure-mgmt-resource==24.0.0
bidict==0.23.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
h11==0.16.0
idna==3.11
isodate==0.7.2
msal==1.34.0
msal-extensions==1.3.1
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
python-engineio==4.13.0
python-socketio==5.16.0
requests==2.32.5
simple-websocket==1.1.0
typing_extensions==4.15.0
uptime-kuma-api==1.2.1
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
